fixed errors, footer edit, cleaned up code, added comments
There was an error with the icons on the contact page so that was fixed, as well as the app section being a bit messed up. Fixed that as well. Added comments so it's easier for everyone to see what goes where, as well as adding in a new footer code that does not mess up the rest of the page. Still needs the right css though.